requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
boto3==1.34.0
botocore==1.34.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
Jinja2==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
cryptography==41.0.8

